bookworm/marketplace/requirements.txt
Flask==2.1.2
grpcio-tools==1.46.3
Jinja2==3.1.2
pytest==7.1.2
